requirements.txt
Django==5.2.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
django-redis==5.4.0
redis==5.2.1
python-decouple==3.8
pillow==11.2.0
django-filter==24.3
drf-spectacular==0.28.0
celery==5.4.0
gunicorn==23.0.0
whitenoise==6.8.2
